+++ interpreter/src/test/scala/pl/touk/nussknacker/engine/spel/SpelExpressionGenSpec.scala
@@ -44,12 +44,12 @@ class SpelExpressionGenSpec
 
   private val NonZeroNumberGen = NumberGen.filterNot(_.doubleValue() == 0)
 
-  private val NoRestrictionsOperarators = List("+", "-", "*")
+  private val NoRestrictionsOperators = List("+", "-", "*")
 
-  private val NotAcceptingZeroOnRrightOperarators = List("/", "%")
+  private val NotAcceptingZeroOnRightOperators = List("/", "%")
 
   test("all combinations of simple arithmetic operators and operands") {
-    val operatorGen = Gen.oneOf(NoRestrictionsOperarators)
+    val operatorGen = Gen.oneOf(NoRestrictionsOperators)
 
     forAll(NumberGen, NumberGen, operatorGen) { (a, b, op) =>
       checkIfEvaluatedClassMatchesExpected(op, a, b)
@@ -73,16 +73,16 @@ class SpelExpressionGenSpec
   }
 
   test("all combinations of operands for divide and modulus operator") {
-    val operatorGen = Gen.oneOf(NotAcceptingZeroOnRrightOperarators)
+    val operatorGen = Gen.oneOf(NotAcceptingZeroOnRightOperators)
 
     forAll(NumberGen, NonZeroNumberGen, operatorGen) { (a, b, op) =>
       checkIfEvaluatedClassMatchesExpected(op, a, b)
     }
   }
 
-  private def checkIfEvaluatedClassMatchesExpected(op: String, a: Any, b: Any) = {
+  private def checkIfEvaluatedClassMatchesExpected(op: String, a: Any, b: Any): Unit = {
     val expr         = s"#a $op #b"
-    val debugMessage = s"expression: $expr; operands: $a, $b; operand types: ${a.getClass}, ${b.getClass}"
+    val debugMessage = s"expression: $expr; operands: $a, $b; operand types: ${a.getClass} and ${b.getClass}."
     logger.debug(debugMessage)
 
     withClue(debugMessage) {
@@ -93,26 +93,27 @@ class SpelExpressionGenSpec
           logger.debug(s"Ignored arithmetic exception: ${a.getMessage}")
         case Failure(other) =>
           fail(other) // shouldn't happen
-        case Success(evaluatedClass) =>
+        case Success(evaluatedValue) =>
           inside(validate(expr, a, b)) {
-            case Valid(TypedExpression(_, TypedObjectWithValue(TypedClass(typedClass, Nil), _), _)) =>
-              typedClass shouldEqual evaluatedClass
+            case Valid(TypedExpression(_, TypedObjectWithValue(TypedClass(typedClass, Nil), typedValue), _)) =>
+              typedValue shouldEqual evaluatedValue
+              typedClass shouldEqual evaluatedValue.getClass
             case Valid(TypedExpression(_, TypedClass(typedClass, Nil), _)) =>
-              typedClass shouldEqual evaluatedClass
+              typedClass shouldEqual evaluatedValue.getClass
             case Valid(TypedExpression(_, TypedUnion(possibleTypes), _)) =>
               val typedClasses = possibleTypes.map(_.asInstanceOf[TypedClass].klass)
-              typedClasses should contain(evaluatedClass)
+              typedClasses should contain(evaluatedValue.getClass)
           }
       }
     }
   }
 
-  private def evaluate(expr: String, a: Any, b: Any): Class[_] = {
+  private def evaluate(expr: String, a: Any, b: Any): AnyRef = {
     val spelParser        = new org.springframework.expression.spel.standard.SpelExpressionParser()
     val evaluationContext = new StandardEvaluationContext()
     evaluationContext.setVariable("a", a)
     evaluationContext.setVariable("b", b)
-    spelParser.parseRaw(expr).getValue(evaluationContext).getClass
+    spelParser.parseRaw(expr).getValue(evaluationContext)
   }
 
   private def validate(expr: String, a: Any, b: Any): ValidatedNel[ExpressionParseError, TypedExpression] = {